pyproject.toml
[tool.poetry]
name = "cron-times"
version = "0.4.0"
description = "Show schdueled jobs in a more readable way"
authors = ["tzing <tzingshih@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "cron_times"}]

[tool.poetry.dependencies]
python = "^3.10"
click = {version = "^8.1.3", optional = true}
colorlog = {version = "^6.7.0", optional = true}
croniter = "^1.3.8"
flask = "^2.2.2"
httpx = {version = "^0.23.3", optional = true}
markdown2 = "^2.4.6"
pypugjs = "^5.9.12"
ruamel-yaml = "^0.17.21"
tzlocal = {version = "^4.2", optional = true}

[tool.poetry.extras]
cli = ["click", "colorlog", "httpx", "tzlocal"]

[tool.poetry.scripts]
"cron-times" = "cron_times.__main__:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
